+++ src/pages/OrderPage/constants/senderSection.ts
@@ -3,6 +3,7 @@ const SENDER_SECTION_CONSTANTS = {
   NAME_PLACEHOLDER: "이름을 입력해주세요",
   NAME_ERROR: "이름을 입력해주세요.",
   INFO_LABEL: "* 실제 선물 발송 시 발신자이름으로 반영되는 정보입니다",
+  DEFAULT_SENDER_NAME: "",
 } as const;
 
 export default SENDER_SECTION_CONSTANTS;
